--- a/src/MonteCarlo.jl
+++ b/src/MonteCarlo.jl
@@ -4,11 +4,13 @@ using Reexport
 @reexport using MonteCarloObservable
 import MonteCarloObservable.AbstractObservable
 using StableDQMC, LightXML, Parameters, Requires
-using JLD, TimerOutputs, LoopVectorization
+using JLD, TimerOutputs
+using LoopVectorization, RecursiveFactorization
 
 using Printf, SparseArrays, LinearAlgebra, Dates, Random
 
 include("helpers.jl")
+include("inplace_udt.jl")
 export enable_benchmarks, disable_benchmarks, print_timer, reset_timer!
 include("flavors/abstract.jl")
 include("models/abstract.jl")