--- a/src/flavors/DQMC/DQMC.jl
+++ b/src/flavors/DQMC/DQMC.jl
@@ -538,7 +538,7 @@ function replay!(
     prepare!(mc.measurements, mc, mc.model)
     for i in mc.last_sweep+1:mc.p.measure_rate:length(configs)
         mc.conf = configs[i]
-        mc.s.greens, mc.s.log_det = calculate_greens_and_logdet(mc, nslices(mc))
+        mc.s.greens, _ = calculate_greens_and_logdet(mc, nslices(mc))
         measure!(mc.measurements, mc, mc.model, i)
         mc.last_sweep = i