--- a/src/flavors/DQMC/stack.jl
+++ b/src/flavors/DQMC/stack.jl
@@ -241,8 +241,8 @@ Updates stack[idx+1] based on stack[idx]
     end
 
     @views rmul!(mc.s.curr_U, Diagonal(mc.s.d_stack[:, idx]))
-    mc.s.u_stack[:, :, idx + 1], mc.s.d_stack[:, idx + 1], T = udt!(mc.s.curr_U)
-    @views vmul!(mc.s.t_stack[:, :, idx + 1],    T, mc.s.t_stack[:, :, idx])
+    @views udt_AVX!(mc.s.u_stack[:, :, idx + 1], mc.s.d_stack[:, idx + 1], mc.s.curr_U)
+    @views vmul!(mc.s.t_stack[:, :, idx + 1], mc.s.curr_U, mc.s.t_stack[:, :, idx])
 end
 """
 Updates stack[idx] based on stack[idx+1]
@@ -255,8 +255,8 @@ Updates stack[idx] based on stack[idx+1]
     end
 
     @views rmul!(mc.s.curr_U, Diagonal(mc.s.d_stack[:, idx + 1]))
-    mc.s.u_stack[:, :, idx], mc.s.d_stack[:, idx], T = udt!(mc.s.curr_U)
-    @views vmul!(mc.s.t_stack[:, :, idx], T, mc.s.t_stack[:, :, idx + 1])
+    @views udt_AVX!(mc.s.u_stack[:, :, idx], mc.s.d_stack[:, idx], mc.s.curr_U)
+    @views vmul!(mc.s.t_stack[:, :, idx], mc.s.curr_U, mc.s.t_stack[:, :, idx + 1])
 end
 
 # Green's function calculation
@@ -265,15 +265,21 @@ Calculates G(slice) using mc.s.Ur,mc.s.Dr,mc.s.Tr=B(slice)' ... B(M)' and
 mc.s.Ul,mc.s.Dl,mc.s.Tl=B(slice-1) ... B(1)
 """
 @bm function calculate_greens(mc::DQMC)
-    # U, D, T = udt(mc.s.greens) after this
-    mc.s.U, mc.s.D, mc.s.T = udt_inv_one_plus(
-        UDT(mc.s.Ul, mc.s.Dl, mc.s.Tl),
-        UDT(mc.s.Ur, mc.s.Dr, mc.s.Tr),
-        tmp = mc.s.U, tmp2 = mc.s.T, tmp3 = mc.s.tmp,
-        internaluse = true
+    # # U, D, T = udt(mc.s.greens) after this
+    # mc.s.U, mc.s.D, mc.s.T = udt_inv_one_plus(
+    #     UDT(mc.s.Ul, mc.s.Dl, mc.s.Tl),
+    #     UDT(mc.s.Ur, mc.s.Dr, mc.s.Tr),
+    #     tmp = mc.s.U, tmp2 = mc.s.T, tmp3 = mc.s.tmp,
+    #     internaluse = true
+    # )
+    # vmul!(mc.s.tmp, mc.s.U, Diagonal(mc.s.D))
+    # vmul!(mc.s.greens, mc.s.tmp, mc.s.T)
+    # mc.s.greens
+    calculate_greens_AVX!(
+        mc.s.Ul, mc.s.Dl, mc.s.Tl,
+        mc.s.Ur, mc.s.Dr, mc.s.Tr,
+        mc.s.greens
     )
-    vmul!(mc.s.tmp, mc.s.U, Diagonal(mc.s.D))
-    vmul!(mc.s.greens, mc.s.tmp, mc.s.T)
     mc.s.greens
 end