For timing/benchmarking I ran:
```julia
using MonteCarlo
model = HubbardModelAttractive(L=6, dims=2)
MonteCarlo.enable_benchmarks()
dqmc = DQMC(model, beta=1.0)
run!(dqmc, sweeps=100, thermalization=100)
MonteCarlo.reset_timer!()

# This run tends to still be a bit slower
dqmc = DQMC(model, beta=1.0)
run!(dqmc, sweeps=1000, thermalization=1000)
MonteCarlo.reset_timer!()

dqmc = DQMC(model, beta=1.0)
run!(dqmc, sweeps=1000, thermalization=1000)
MonteCarlo.print_timer()
```

## On master:

```
─────────────────────────────────────────────────────────────────────────────────────────────────────
                                                             Time                   Allocations      
                                                     ──────────────────────   ───────────────────────
                  Tot / % measured:                       14.4s / 62.4%           1.70GiB / 95.9%    

Section                                      ncalls     time   %tot     avg     alloc   %tot      avg
─────────────────────────────────────────────────────────────────────────────────────────────────────
run!                                              1    8.99s   100%   8.99s   1.64GiB  100%   1.64GiB
  propagate                                   40.0k    6.47s  72.0%   162μs   1.63GiB  100%   42.7KiB
    propagate                                 4.00k    3.19s  35.4%   797μs   1.22GiB  74.9%   321KiB
      calculate_greens                        4.00k    2.24s  24.9%   559μs   0.98GiB  60.1%   258KiB
      calculate_logdet                        4.00k    844ms  9.40%   211μs   84.5MiB  5.05%  21.6KiB
    add_slice_sequence_right                  2.00k    1.56s  17.3%   778μs    206MiB  12.3%   106KiB
    add_slice_sequence_left                   2.00k    1.27s  14.1%   635μs    206MiB  12.3%   105KiB
    calculate_greens                              1    961μs  0.01%   961μs    258KiB  0.02%   258KiB
    calculate_logdet                              1   67.4μs  0.00%  67.4μs   21.6KiB  0.00%  21.6KiB
  sweep_spatial                               40.0k    2.29s  25.5%  57.3μs   1.66KiB  0.00%    0.04B
    accept_local!                             1.32M    1.63s  18.1%  1.23μs     0.00B  0.00%    0.00B
    propose_local                             1.44M    221ms  2.46%   154ns     0.00B  0.00%    0.00B
  measure!                                      300   2.99ms  0.03%  10.0μs   7.83MiB  0.47%  26.7KiB
  add_slice_sequence_left                         1    284μs  0.00%   284μs    108KiB  0.01%   108KiB
─────────────────────────────────────────────────────────────────────────────────────────────────────
```


## After removing logdet

```
─────────────────────────────────────────────────────────────────────────────────────────────────────
                                                             Time                   Allocations      
                                                     ──────────────────────   ───────────────────────
                  Tot / % measured:                       15.0s / 44.5%           1.57GiB / 99.0%    

Section                                      ncalls     time   %tot     avg     alloc   %tot      avg
─────────────────────────────────────────────────────────────────────────────────────────────────────
run!                                              1    6.67s   100%   6.67s   1.55GiB  100%   1.55GiB
  propagate                                   40.0k    4.25s  63.7%   106μs   1.54GiB  99.5%  40.5KiB
    propagate                                 4.00k    1.80s  27.0%   450μs   1.14GiB  73.6%   299KiB
      calculate_greens                        4.00k    1.70s  25.4%   424μs   0.98GiB  63.3%   258KiB
    add_slice_sequence_left                   2.00k    1.08s  16.2%   540μs    206MiB  12.9%   105KiB
    add_slice_sequence_right                  2.00k    887ms  13.3%   444μs    206MiB  13.0%   106KiB
    calculate_greens                              1   1.38ms  0.02%  1.38ms    258KiB  0.02%   258KiB
  sweep_spatial                               40.0k    2.23s  33.4%  55.7μs   1.66KiB  0.00%    0.04B
    accept_local!                             1.32M    1.61s  24.2%  1.22μs     0.00B  0.00%    0.00B
    propose_local                             1.44M    204ms  3.05%   141ns     0.00B  0.00%    0.00B
  measure!                                      300   2.88ms  0.04%  9.62μs   7.83MiB  0.49%  26.7KiB
  add_slice_sequence_left                         1    665μs  0.01%   665μs    108KiB  0.01%   108KiB
─────────────────────────────────────────────────────────────────────────────────────────────────────
```


## logdet removal + UDT optimizations

```
 ─────────────────────────────────────────────────────────────────────────────────────────────────────
                                                              Time                   Allocations      
                                                      ──────────────────────   ───────────────────────
                   Tot / % measured:                       3.88s / 81.7%            436MiB / 81.9%    

 Section                                      ncalls     time   %tot     avg     alloc   %tot      avg
 ─────────────────────────────────────────────────────────────────────────────────────────────────────
 run!                                              1    3.17s   100%   3.17s    357MiB  100%    357MiB
   sweep_spatial                               40.0k    1.91s  60.3%  47.7μs   1.66KiB  0.00%    0.04B
     accept_local!                             1.32M    1.38s  43.6%  1.04μs     0.00B  0.00%    0.00B
     propose_local                             1.44M    180ms  5.70%   125ns     0.00B  0.00%    0.00B
   propagate                                   40.0k    1.22s  38.4%  30.4μs    348MiB  97.7%  8.92KiB
     propagate                                 4.00k    443ms  14.0%   111μs    305MiB  85.7%  78.2KiB
       calculate_greens                        4.00k    346ms  10.9%  86.6μs    143MiB  40.0%  36.5KiB
     add_slice_sequence_right                  2.00k    183ms  5.77%  91.4μs   21.6MiB  6.05%  11.0KiB
     add_slice_sequence_left                   2.00k    171ms  5.40%  85.4μs   21.2MiB  5.95%  10.9KiB
     calculate_greens                              1    101μs  0.00%   101μs   41.5KiB  0.01%  41.5KiB
   measure!                                      300   5.43ms  0.17%  18.1μs   7.83MiB  2.20%  26.7KiB
   add_slice_sequence_left                         1   87.9μs  0.00%  87.9μs   13.3KiB  0.00%  13.3KiB
─────────────────────────────────────────────────────────────────────────────────────────────────────
```